+++ src/100-bindings/typescript/writeStruct.ts
