+++ src/100-bindings/writeTypescript.ts
